requirements.txt
pandas
requests
apache-airflow
pytest